+++ src/components/ArcGISWebMap.tsx
@@ -22,6 +22,7 @@ import useIsPageVisible from '../hooks/useIsPageVisible'
 import { ArcGISWebMapElementValue } from '@oneblink/types/typescript/arcgis'
 import { FormElementValueChangeHandler } from '../types/form'
 import '../styles/arcgis-external.css'
+import { Box, Divider, IconButton } from '@mui/material'
 
 type Props = {
   element: FormTypes.ArcGISWebMapElement
@@ -30,6 +31,56 @@ type Props = {
   onChange: FormElementValueChangeHandler<ArcGISWebMapElementValue>
   'aria-describedby'?: string
   setIsDirty: () => void
+}
+
+type SketchCreateTool = Parameters<Sketch['create']>[0]
+
+function DrawingOptionsList({
+  options,
+  sketchTool,
+  onClose,
+  sketchToolType,
+  setSelectedGraphicAttributes,
+}: {
+  options: { id: string; label: string; value: string }[]
+  onClose: () => void
+  sketchTool: Sketch
+  sketchToolType: SketchCreateTool
+  setSelectedGraphicAttributes: (opt: { label: string; value: string }) => void
+}) {
+  return (
+    <div className="esri-widget">
+      <Box sx={{ display: 'flex', alignItems: 'center' }}>
+        <b style={{ padding: '0.5rem' }}>Select an option</b>
+        <IconButton
+          onClick={(e) => {
+            e.stopPropagation()
+            onClose()
+          }}
+          data-cypress="copy-to-clip-board-button"
+        >
+          <MaterialIcon fontSize="small">close</MaterialIcon>
+        </IconButton>
+      </Box>
+      <Divider sx={{ backgroundColor: 'unset', margin: '0px 8px' }} />
+      <Box sx={{ maxHeight: '200px', overflowY: 'auto' }}>
+        {options?.map(({ id, value, label }) => (
+          <div
+            key={id}
+            className="ob-list__item is-clickable"
+            style={{ alignItems: 'center' }}
+            onClick={() => {
+              onClose()
+              sketchTool?.create(sketchToolType)
+              setSelectedGraphicAttributes({ value, label })
+            }}
+          >
+            {label}
+          </div>
+        ))}
+      </Box>
+    </div>
+  )
 }
 
 function FormElementArcGISWebMap({
@@ -52,6 +103,14 @@ function FormElementArcGISWebMap({
   const [loadError, setLoadError] = React.useState<Error>()
   const [isLoading, setIsLoading] = React.useState<boolean>(false)
   const isPageVisible = useIsPageVisible()
+  const [selectedGraphicAttributes, setSelectedGraphicAttributes] =
+    React.useState<{
+      value: string
+      label: string
+    }>()
+  // only used when an allowed drawing tool also has a list of graphic attribute options to display to the user
+  const [activeSketchToolMenu, setActiveSketchToolMenu] =
+    React.useState<SketchCreateTool>()
 
   const updateDrawingInputSubmissionValue = React.useCallback(() => {
     const updatedGraphics = drawingLayerRef.current?.graphics
@@ -96,7 +155,17 @@ function FormElementArcGISWebMap({
       'create',
       (sketchEvent) => {
         if (sketchEvent.state === 'complete') {
+          if (selectedGraphicAttributes) {
+            sketchEvent.graphic.attributes = {
+              name: selectedGraphicAttributes.value,
+              label: selectedGraphicAttributes.label,
+            }
+            setSelectedGraphicAttributes(undefined)
+          }
           updateDrawingInputSubmissionValue()
+        }
+        if (sketchEvent.state === 'cancel') {
+          setSelectedGraphicAttributes(undefined)
         }
       },
     )
@@ -123,6 +192,7 @@ function FormElementArcGISWebMap({
     )
 
     const deleteListener = sketchToolRef.current?.on('delete', () => {
+      mapViewRef.current?.closePopup()
       updateDrawingInputSubmissionValue()
     })
 
@@ -142,18 +212,70 @@ function FormElementArcGISWebMap({
       },
     )
 
+    const sketchToolListener = sketchToolRef.current?.viewModel.watch(
+      'activeTool',
+      () => {
+        if (selectedGraphicAttributes) {
+          return
+        }
+        const activeTool = sketchToolRef.current?.activeTool
+        const hasGraphicAttributeOptions = !!element.allowedDrawingTools?.find(
+          (tool) => tool.type === activeTool,
+        )?.graphicAttributeOptions?.length
+
+        if (activeTool) {
+          if (
+            hasGraphicAttributeOptions &&
+            (activeTool === 'point' ||
+              activeTool === 'polygon' ||
+              activeTool === 'circle' ||
+              activeTool === 'rectangle' ||
+              activeTool === 'polyline')
+          ) {
+            // the sketch.create() fn only allows active tools of the above 5 types
+            // hence we must check above to keep the types happy further down the line
+            sketchToolRef.current?.cancel()
+            setActiveSketchToolMenu(activeTool)
+          } else {
+            setActiveSketchToolMenu(undefined)
+          }
+        }
+      },
+    )
+
+    mapViewRef.current?.on('click', (event) => {
+      mapViewRef.current?.hitTest(event).then((response) => {
+        // the "hit test" will typically yield the graphic we want to display the popup for,
+        // and an ArcGIS built-in graphic which highlights the selected graphic.
+        // By filtering for the graphic with the "label" attribute, we're able to reliably
+        // get the graphic we want to display the popup for.
+        const result = response.results
+          .filter((r) => r.type === 'graphic')
+          .find((r) => !!r.graphic.attributes.label)
+
+        if (result) {
+          mapViewRef.current?.openPopup({
+            location: result.mapPoint,
+            title: result.graphic.attributes.label,
+          })
+        }
+      })
+    })
+
     return () => {
       createListener?.remove()
       updateListener?.remove()
       deleteListener?.remove()
       mapViewChangeListener?.remove()
+      sketchToolListener?.remove()
     }
   }, [
     isLoading,
     value,
     updateDrawingInputSubmissionValue,
     updateMapViewSubmissionValue,
-    element.readOnly,
+    element,
+    selectedGraphicAttributes,
   ])
 
   const onSubmissionValueChange = React.useCallback(() => {
@@ -310,6 +432,7 @@ function FormElementArcGISWebMap({
           })
           sketchToolRef.current = sketch
           view.ui.add(sketch, 'bottom-right')
+          view.ui.add('OneBlinkDrawingOptionsSelector', 'bottom-right')
         }
 
         // once the view has loaded
@@ -373,6 +496,23 @@ function FormElementArcGISWebMap({
         id={id}
         aria-describedby={props['aria-describedby']}
       />
+      <div id="OneBlinkDrawingOptionsSelector">
+        {!!activeSketchToolMenu && sketchToolRef.current && (
+          <DrawingOptionsList
+            options={
+              element.allowedDrawingTools?.find(
+                (tool) => tool.type === activeSketchToolMenu,
+              )?.graphicAttributeOptions || []
+            }
+            onClose={() => setActiveSketchToolMenu(undefined)}
+            setSelectedGraphicAttributes={(opt) => {
+              setSelectedGraphicAttributes(opt)
+            }}
+            sketchToolType={activeSketchToolMenu}
+            sketchTool={sketchToolRef.current}
+          />
+        )}
+      </div>
     </>
   )
 }